requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
boto==2.32.1
itsdangerous==0.24
py-bcrypt==0.4
wsgiref==0.1.2
